Changed route prefix (temporary)
Ultimately, the editor needs to respond to requests going to `/Edit`. This is not currently possible, however, due to a bug in `TopicRoutingService`, which is unaware of prefixes or namespaces, and assumes the entire path is the Topic path. Until this is fixed, the route is being set to `Web` to continue development.