Migrated to new `IModelBinder` interface
The legacy `DefaultModelBinder` base class—and its `CreateModel()` method—is no longer supported. That said, a very similar `IModelBinder` interface—with a `BindModelAsync()` method—is supported. This update migrates to the new *ASP.NET Core* format. Note that this may require additional annotations or an `IModelBinderProvider` to ensure that this is honored during the binding process; that will be determined as part of integration testing.