Remove unnecessary `JsonNetResult`
In the *ASP.NET Framework*, C# objects were serialized to JSON objects using the same casing—i.e., PascalCase for properties. At best, this was inconsistent with standards for JSON, which typically used camelCase properties; at worse, it introduced compatibility issues with reusable libraries. To mitigate this, a custom `JsonNetResult` class was created that was capable of serializing the class names as camelCase using the `NewtonSoft.Json` library. In *ASP.NET Core*, however, camelCase is now implemented by default during the JSON serialization, and so this shouldn't be necessary.