Updated `csproj` file format
The legacy `csproj` file format required a lot of boilerplate code, such as build targets, that are now assumed in *ASP.NET Core*. It no longer requires that every file be explicitly enumerated—all files are automatically imported by default. And it does away with the separate `packages.config` file, instead integrating it with the `csproj`. As part of this, package references have been updated to their *.NET Core* versions, including the updating of the `Ignia.Topics` library to `3.8.0-alpha.255`, and the introduction of `Microsoft.AspNetCore.App`. Finally, the project type is updated to a `netcoreapp2.2` target, obviously.