Migrated routing format
In the legacy *ASP.NET Framework* version, the area was defined by a `AreaRegistration` class (`EditorAreaRegistration`) and a `[RouteArea()]` attribute; the default route was configured via a `RouteConfig` class. In *ASP.NET Core*, this is now all done via the `UseMvc()` configuration method in the `Statup` class, using `UseMvcWithDefaultRoute() and `MapAreaRoute()`. This is a much simpler format—though we'll need to confirm it correctly recognizes it under `/OnTopic/`, instead of e.g. `/Areas/Editor/` or `/Editor/`.