Replaced `ControllerFactory` with `Activator`
In *ASP.NET Core*, the _Composition Root_ is now in an `IControllerActivator` (and `IViewComponentActivator`) instead of a `IControllerFactory`. This is updated to use the standard activator format for *OnTopic* and registered in the `Startup` class. As per Ignia's guidance, the same class and instance is used for both controllers and view components, thus _singleton_ dependencies to be shared between them.