Established user secret
The user secret allows secrets to be stored on a user-secured location on the local drive instead of in the project file, where it will be committed to git. On a production server, this can instead use e.g. an environmental variable or, on *Azure*, a server-side connection string (since we're currently using it for connection strings).