Modified to use view models
Previously, the `EditorViewModel` returned entity objects (e.g., `Topic` and `ContentTypeDescriptor`) as well as services (e.g., `ITopicRepository`). This is not a best practice. The new version is setup to instead return `TopicViewModel` and `ContentTypeDescriptorTopicViewModel`, which are limited exclusively to the properties required by the interface. The `ITopicRepository` is no longer returned; other approaches should be discovered for addressing this (such as `ViewComponent`s).