Updated `AttributeViewModel` to work with `AttributeTypeViewComponent`
With the migration from `@Html.EditorFor()` to `AttributeTypeViewComponentBase`, the needs of the `AttributeViewModel` have changed. It now simply needs to reference the `AttributeDescriptorTopicViewModel` as well as provide a reference for the `InheritedValue` and `Value` properties. It doesn't need to relay the original `AttributeDescriptor` entity, nor the entirety of the current `Topic` entity. And it certainly doesn't need access to services, such as the `ITopicRepository`.