Manually injected attributes
The current `TopicMappingService` doesn't support mapping an arbitrary collection of topics (i.e., a reference collection). Only collections originating as children, relationships, or nested topics are supported, unless the return type is derivable from the target. Since that isn't the case with the `AttributeDescriptors` collection, these must be manually mapped.