Introduced `DefaultAttributeTypeViewComponent`
Unlike the `AttributeTypeViewComponentBase`, the `DefaultAttributeTypeViewComponent` has a default implementation of the `InvokeAsync()` method, so that implementor need not recreate it. This is useful given that the majority of attribute types will return an `AttributeViewModel` with standard options. For those that need to customize how the `AttributeViewModel` is created, or return a derivative view model, the `AttributeTypeViewComponentBase` class should still be used.