Replaced `for` with `foreach`
A `foreach()` statement is a bit cleaner and easier to read than a `for()` statement. The `for()` statement had been used to ensure a unique `Id` for each `Attribute`. Since that's now using the separate `index` variable to ensure uniqueness across _Display Groups_, however, the `for()` no longer provides value.