Added defensive code to model binder
The model binder in *ASP.NET Core* will continue to iterate over a collection until a `null` result is returned. This was previously causing a null reference exception. The added defensive code mitigates that. In addition, I updated the commenting and formatting since the method had become a bit more involved.