Moved `Key` and `EditorType` to `AttributeViewModel`
`Key` and `EditorType` used to be on `AttributeViewModel` but were removed because they could easily be pulled from the `AttributeDescriptor` property. This, however, introduced complexities with the binding model, since the *ASP.NET Core* generated HTML names now included `AttributeDescriptor` in them. Adding these back to `AttributeViewModel` not only allows these to map cleanly to the interface of the `EditorBindingModel`, but also provides a more elegant place to remove the legacy `*.ascx` extension, if it exists. (Long term, that needs to be removed at the data level, but in the short-term we're providing this as a point of backward compatibility for legacy data sets.)