Manually set `Key` and `Value`
For reasons that aren't clear, `Key` and `Value` aren't getting implicitly set by *ASP.NET Core*. The reasoning is unclear, and may represent a more fundamental bug in the implementation. As a workaround, these values are being set as part of the model binder. This works for our needs, but isn't ideal. This may be worth further investigation at a later date.