Moved static resources to `/wwwroot`
In *ASP.NET Core*, static files _most_ be in the `wwwroot` folder; they can't be called from e.g. the `Views` or `Areas` folders, as they could in *ASP.NET MVC Framework*. (Or, rather, they can, but it requires custom configuration, and may make redistribution more involved.) To mitigate this, the `/Areas/Editor/Shared/` folder is being moved to the `/wwwroot/Shared/` folder. This isn't ideal, and we may need to revisit this later as part of setting up the editor for distribution. But it at least solves the immediate issue of the interface not rendering correctly.