Migrated to RCL/Host format
Introduced a new *ASP.NET Core* website project, `Ignia.Topics.Editor.Mvc.Host`, that acts as a sample host application. Then migrated the `Ignia.Topics.Editor.Mvc` to a _Razor Class Library (RCL)_, so that it can be embdded into the `Host` project—or redistributed to any other *ASP.NET Core* project. As part of that, migrated files relates to the _Development Context_ to the `Host` project (e.g., `Program`, `Startup`, and `SampleActivator`).