Moved model binding to `Mvc` project
In order to allow the `Models` library to be focused exclusively on models—without any *ASP.NET Core* dependencies—I moved the `EditorAttributeModelBinder` to `Ignia.Topics.Editor.Infrastructure` under the `Mvc` project. Since this mandated removing the `[ModelBinder()]` attribute from `EditorAttribute`, this also meant replacing that with a new `EditorAttributeModelBinderProvider`, as well as an `IMvcBuilder` extension for configuring it.