Converted `Models` library to *.NET Standard*
With the *ASP.NET Core* dependencies on the model binder removed, the `Models` library can be converted to a standard *.NET Standard* library. As a result, it no longer needs any of the *ASP.NET Core* dependencies—nor, in fact, does it need any of the ependencies on several other libraries inherited from the original version (e.g., `Microsoft.Net.Compilers`, `Microsoft.ApplicationInsights`, or `Ignia.Topics.Data.Caching`, for instance). Those are removed. Finally, the `app.config` is also removed, since it is no longer needed or even used. Basically, this is now a really simple set of POCO classes now.