Introduced `TopicLookupAttributeViewModel`
The `TopicLookupViewComponent` requires a custom view model in order to relay the select box options. This can be accomplished by simply adding a `List<SelectListItem>` to a derivation of `AttributeViewModel`. That said, the `SelectListItem` class is in an assembly that drags with it the majority of *ASP.NET Core*—none of which the rest of the models classes need, or should need. As a result, it is being created in the `Mvc` project. This is a bit inconsistent, but is preferrable to the `Models` project needing to take on a lot of new dependencies for one class.