Fix `Configuration` return type
The return type of the `Configuration` property was set to `Dictionary<string, string>`. But the `AttributeDescriptor` it is mapping to uses `IDictionary<string, string>`. As such, the `TopicMappingService` isn't able to do a 1:1 transfer of the existing object. Migrating to `IDictionary<>` resolves this, and ensures that the `Configuration` property gets populated during mapping.