Removed redundant properties
In the .NET Core project format, project properties are now in the `csproj` file, not in the `AssemblyInfo.cs` file. For some reason, this worked fine up until upgrading to the latest version of the `Ignia.Topics.dll`. With the latest version, however, the `dotnet` compiler started tripping on the duplicate properties. Given that, removed all duplicates, both allowing the editor project to compile, and making it consistent with other .NET Core projects.