Establish file monitoring
Unfortunately, edits to Razor views don't automatically trigger recompilation of Razor Class Libraries (RCLs) like they do in standard MVC Applications. That means that modifying views requires recompiling and reloading the application—a slow process. This can't be entirely avoided, but this workaround at least automates and streamlines the process to be triggered by the `RazorViewEngine` automatically whenever files are modified without needing to restart the server. This still requires that the assembly be recompiled and that the application reload (including downloading of the database), so it may take upwards of 30 seconds before a change in a `*.cshtml` file will be available. Better tan several minutes, though?