Remove unnecessary dependencies
Like *Application Insights*, these dependencies are useful for most projects, but not strictly necessary for the barebones purposes of this project. Namely, we don't need the ability to generate new e.g. controllers, nor do we need to instrument any logging in the current solution. These can be added later if they prove necessary—but, more likely, they'll only be necessary in client projects.