Rename `Resources` back to `wwwroot`
In **ASP.NET Core 2.2**, there is an extension that allows static resources (assets) to be embedded. As part of that, a best practice recommended by some—and which **Ignia** adopted—is to rename that folder from `wwwroot` to `Resources` to better indicate that this isn't a website, but that these are, in fact, static resources that will be merged with a website. In **ASP.NET Core 3.0**, embedding static resources is now built into **Razor Class Libraries** (woohoo!)—and the preferred/OOTB convention is to use `wwwroot`. Given that, and in preparation for the migration to **ASP.NET Core 3.0**, we're renaming this _back_ to `wwwroot`. We may reevaluate this later, but for now we want to establish a consistent baseline.