Updated to endpoint routing
Updated from the legacy `UseMvc()` route provider to the new(er) `UseEndpoints()` routing. This includes use of the new `MapAreaControllerRoute()`, as well as the use of `MapControllers()` in place of the legacy `UseMvcWithDefaultRoute()` configuration. Again, this isn't strictly necessary, but maintains consistency with out-of-the-box **ASP.NET Core 3** templates and best practices.