Removed `UseCookiePolicy()`
We don't use the `UseCookiePolicy()` for the purpose of the host site, but it was there for familiarity and consistency with the basic **ASP.NET Core 2.2** template that the editor would likely be incorporated into. This is no longer used by default in the new **ASP.NET Core 3.0** templates, however, so it is being removed.