Updated new *Razor Class Library* static file format
In **ASP.NET Core 2.0**, static assets could be embedded as resources into a **Razor Class Library**. This was a really welcome feature, but required quite a bit of scaffolding—and, thus, the previous introduction of the `AddTopicEditor()` configuration extension, and its associatd `EditorConfigureOptions` class. In **ASP.NET Core 3**, static resources are embedded by default—so long as they're in the `wwwroot` folder—and accessibly to the referencing site without any special wiring up of e.g. `ContentTypeProvider`s, such as the previous `FileExtensionContentTypeProvider`. This greatly simplifies how the `Startup` class works, and eliminates the need for injecting the `IHostingEnvironment` service (which is convenient, as it's now obsolete). That said, this does introduce some changes to the legacy approach. Namely, files are no longer mapped onto the parent site's `wwwroot` folder, as they previously were; instead, they are now placed in a special `/_content` folder, namespaced by the project name. As a result, the library references in the `_Layout.cshtml` are also updated as part of this.