Introduced `AttributeViewModel<T>`
The `AttributeViewModel<T>` accepts an options class as the generic type, which can be `DefaultOptions` or any derivative. This allows it to introduce a strongly type `Options` property. Once implemented, this will give all views access to the `Options` used to initialize the attribute view component. This prevents the need to duplicate these properties on the `AttributeViewModel` (or derivative) in the case that they need to be relayed directly to the view.