Introduced `CssClass` and `IsEnabled` passthrough properties
These properties were already exposed to `DefaultOptions`. Shared Razor pages that might need access to these, however, such as `_Layout.cshtml`, can't access the `Options`. This is because a strongly typed `AttributeViewModel<T>` can't be cast to e.g. `AttributeViewModel<DefaultOptions>` unless it is explicitly uses `DefaultOptions`. As such, defining this on `AttributeViewModel`, but then overriding them on `AttributeViewModel<T>`, allows them to a) be accessed via shared pages (which can be bound to `AttributeViewModel`), while b) relying on `AttributeViewModel<T>` (which derives from `AttributeViewModel`) to access those values directly from `Options`. It's not an especially elegant work around since it requires repeating these properties three times (in `DefaultOptions`, `AttributeViewModel`, and `AttributeViewModel<T>`. Fortunately, we don't expect there to be many shared options that need to be relayed down to the views _and_ accessed by shared partials, so this shouldn't happen often. For most other configuration properties, they will only be access by views strongly typed to their specific view model. Implemented a call to this in `_Layout.cshtml` as a proof-of-concept.