Renamed `Options` to `TopicList`
The `TopicLookupViewModel` implemented an `Options` property for tracking a list of topics to be bound to a dropdown list. This name conflicted with the new `Options` property, which was added to the `AttributeViewModel` base in order to access the options object. To eliminate this conflict, I renamed `Options` to `TopicList`. This is actually a better name as a) it ties it into our legacy name for this collection, and b) it's more descriptive of what _type_ of list it represents.