Added debug data to `_Layout.cshtml`
To simplify common debugging during development, am writing out the underlying `Value` and `InheritedValue` values directly to the interface in a hidden `<div />`. We actually did this in the legacy editor in many of our controls; having this centralized avoids the need for repetition. Technically, this data can be gotten through formal debugging, but this provides a quick way of verifying the most commonly required data during development.