Establish blank `VersionHistory`
The `TopicMappingService` will automatically create and populate the `VersionHistory` property _if_ a version history exists. If there aren't any versions, however, it won't get created—which means views need to search for nulls before e.g. iterating over the collection. To simplify this, we're now creating a new `List<DateTime>` as a default value so that it's never null.