Introduced `BooleanAttributeViewModel`
Introduced a `BooleanAttributeViewModel` which exposes a new `IsSelected()` method for determining if the current attribute is selected. This is necessary since the value isn't usually in a format that the `@Html.RadioButtonFor()` method understands (i.e., it doesn't understand `1` and `0` instead of `true` and `false`). Exposed this on the view in order to conditionally set the checked state of each radio button. Also introduced a `BooleanOptions` object—though, at this point, there aren't actually any boolean options, so it's just inheriting the `DefaultOptions`.