Established deprecated pass-through for `FileViewComponent`
In a previous commit (#), we renamed `FileViewComponent` to `FileListViewComponent`. Legacy databases, however, may still be calling `FileViewComponent`. These should be updated. To simplify this transition, however, an instance of `FileViewComponent` was reintroduced, which simply subclasses `FileListViewComponent`, thus ensuring it isn't duplicating any logic. The class is marked as `[Obsolete()]` since it isn't intended to be used—though, users will only discover this in their composition root, since no one is instantiating this directly.