Implemented initial version of `RelationshipsViewComponent`
This includes a new `RelationshipsOptions` class which exposes all of the `DefaultConfiguration` properties that the relationships attribute will accept. This provides a convenient way of exposing and documenting the available options in a simple format—though, of course, those options still aren't exposed to the user interface, unfortunately. This implementation includes the **ExtJS** `TreeView` control, which is configured inline, and relays the `Model.Options` to the `dataUrl` (which pulls values from the `EditorController.JsonAsync()` action). Ideally, we'd subclass the script so that common attributes aren't repeated for each relationship on a page; ultimately, however, we want to replace the **ExtJS** implementation with a lighter-weight, stand-alone tree view control. As a result, we're not worrying about optimizing this since it isn't something we want to maintain for too long.