Updated `TopicLookupViewComponent` to use `TopicLookupOptions`
The new options object approach provides a more elegant way of handling large amounts of parameters, which can be easily relayed down to the view. This also greatly simplifies the implementation of `TopicLookupviewComponent` by eliminating the need for a couple of the legacy local functions (`setBooleanConfiguration()` and `setConfiguration()`) which are now handled directly by the `AttributeDescriptorTopicViewModel`. The `TopicLookupViewComponent` successfully binds to a dropdown list. That said, this is not yet fully implemented; it still needs to account for, specifically, the `TargetUrl` and `TargetPopup` properties, which are needed for its use in `TopicListViewComponent` and in the main `Edit.cshtml`. In both cases, these are used for displaying a list of `ContentTypeDescriptor`s, and allowing a new `Topic` to be created—either by redirecting to a new page (`TargetUrl`) or by spawning a modal window (`TargetPopup`). For the latter scenario, it also needs to implement `OnClientClose`, which provides functionality for how to handle events when a modal window that it spawned is closed. For now, however, this provides a baseline implementation for standard attribute lookups.