Implemented initial version of `TopicListViewComponent`
This version implements a `TopicListOptions` which exposes all configuration properties available via `DefaultConfiguration`, a `TopicListAttributeViewModel` which exposes `IsNew`, `UniqueKey`, and `WebPath` properties, and introduces an initial version of the **ExtJS** `TreeView` control in the view. That said, this is not yet complete: It still needs to expose support for event listeners (`click()`, `startDrag()`, `nodedragover()`, and `movenode()`; the modal popup, via `Options.TargetPopup`; and, finally, the embedded `TopicLookupViewComponent` so that new topics can be added to the list (also by spawning a child modal window). Still, at this point, it at least supports displaying a read-only list of nested topics on the page.