Introduced initial version of `FileListViewComponent`
Introduced a new `FileListOptions` class to establish the properties available via `DefaultConfiguration`, a `FileListAttributeViewModel` to expose the files list, a `GetFiles()` helper function in `FileListViewComponent` in source the files list (based on the `FileListOptions`), and, finally, a `FileList` view (as well as a `File` view for backwards compatibility) to render the files list as a dropdown list. This also required introducing the `IWebHostEnvironment` as a dependency for the `FileListViewComponent` so that it could map local paths; that, in turn, required that dependency being established in `Startup`, passed to `SampleActivator`, and then, finally, related to the `FileViewComponent` and `FileListViewComponent` registrations.