Include dependency scripts in master layout
Previously, scripts were registered directly by the **ASP.NET** _user controls_ that required them, by using the `ScriptManager`. Unfortunately, that functionality isn't supported with **ASP.NET Core** _partial scripts_. Given that, we can either include them in the plugins—which means they won't be in the header or footer, and may be duplicated—or we can just put them directly in the `_Layout.cshtml`, which will result in them being included on every page, even if they're not needed. Neither of these are great options, but the scripts are small, so it's not a huge issue to include them on every page, and especially since they'll be needed sooner or later (i.e., they'll end up cached regardless). Eventually, if we stick with this approach, we'll want to concatenate and minimize these, but that will wait until we've finished setting up a new build process, which will also source these from **npm**, instead of maintaining third-party libraries as part of our repository.