Introduced initial version of `TokenizedTopicListViewComponent`
The new implementation includes a `TokenizedTopicListOptions`, which includes all settable properties from `DefaultConfiguration`; a `TokenizedTopicListAttributeViewModel`, which exposes a JSON-formatted list of previously selected topics; the view component itself, of course, which includes a `GetSelectedTopics()` and `GetTopicJson()` method for looking up the existing value and converting it to JSON; and, of course, the view, which uses the `TokenizedTopicListOptions` to initialize a new instance of the `TokenizedTopics` JavaScript class. Finally, it also includes a minor update to the `TokenizedTopicList.js` file itself to correct the path to the JSON feed. This did expose issues with the `/OnTopic/JSON/` filtering, which need to be addressed, but as those are independent of the `TokenizedTopicListViewComponent`, they'll be deferred to a separate, future commit. In addition, we still need to implement the `IsAutoPostBack` functionality.