Added nullability hint for `GetAttributeViewModel()`
Technically, since we're not (yet) enabling **C# 8.0**'s new nullability annotation context yet, this isn't strictly necessary. Since we have implemented nullability on a few view models, however, it at least helps with validation of calls depending on `GetAttributeViewModel()`. At some point, we'll likely choose to fully implement this, but that's not an immediate priority.