Implemented `Delete()` functionality
Rewrote the **jQuery** statement for handling the delete functionality. There were previously two legacy versions—one which did an AJAX call, but didn't function, and another which confirmed the user's intention, but didn't submit. The new version confirms the user's intention and then, if they confirm, does a redirect. I setup `Delete()` to support `[Get]` for now, so that a simple redirect can be done. In the future, we likely want to update this (and other confirmations) with a fancier confirmation box from either **jQuery** or **Bootstrap** (or, **Foundation**, if we migrate to it), and ideally an AJAX submission. We can reevaluate all of that later; for now, the objective is to maintain the legacy functionality.