Scaffolded `TopicLookup` for navigation
Added a nested `TopicLookup` for listing available content types. As with #1f6d9fc, there are some issues with this still since the `TargetUrl` feature isn't yet implemented. It also incorrectly displays the dropdown box even if there's only a single entry. Plus it shows the label (from `_layout.cshtml`), which isn't desirable. Those issues will need to be worked out. For now, however, this establishes the basic functionality, and provides a basis to add and test additional features. Aside, it's unclear why this manual mapping of `PermittedContentTypes` is necessary; this suggests a bug in the `DynamicTopicViewModelLookupService`; that will require further investigation.