Worked around mapping issue with `ContentTypeDescriptorTopicViewModel`
Previously, `ITopicMappingService.MapAsync<ContentTypeDescriptorTopicViewModel>` would map top-level properties, but run into issues recursing over nested topics and relationships due to the `Ignia.Topics.Editor.Models` assembly not being loaded in the `AppDomain` of the `Ignia.Topics` assembly—even though it was loaded in context of the `Ignia.Topics.Editor.Mvc` assembly which was calling `ITopicMappingService<>`. To work around this, an `EditorViewModelLookupService` was established which provides a static reference to local view models. On the downside, this has to be manually maintained any time a new type is added. On the upside, this solves the problem and is much faster.