Moved `SCSS` files out of `wwwroot`
`SCSS` files need to be compiled, and should not be accessible directly to web clients. Placing them in `wwwroot` was a quick shortcut due to the original compilation process outputting the resulting `CSS` files in the same location. We'll be updating our build process, however, to account for **ASP.NET Core**'s convention of using `wwwroot` as a build/output folder for static files. In preparation for that, these files are being moved back to their original location.