Moved `JS` files out of `wwwroot`
Unlike the `SCSS` files (see #9d71cb2), the `JS` files _can_ be used in situ without relying on a build process to compile and relocate them. _That said_, it is _preferred_ that they, at minimum, be uglified and concactenated into a single file so that we don't need to open several HTTP sessions to download them individually. And, of course, we'd ultimately like to migrate these over to *TypeScript*, at which point they _will_ need a precompilation process. As such, they're being moved _back_ to the root `/Shared` folder, and will be built using the forthcoming **Gulp** configuration.