Deleted unused `CSS` files
The `Site.css` file represents a rollup of vendor `CSS` files from a legacy build process. Currently, however, all `CSS` files are being accessed by their individual vendor files, and the `Site.css` is neither maintained nor referenced. We'll likely be building a new version of this as part of the forthcoming **Gulp** workflow, but there's no need to keep this legacy file around in the interim, as it only causes confusion. Likewise, the `Stylesheet.css` is not used and, in fact, only contains commented out styles.