Introduced new *Gulp.js* workflow
The new workflow compiles, minimizes, and maps the `SCSS` files from `/Shared/Styles` and outputs the results to `/wwwroot/Shared/Styles`. It also validates (`JSHint`), uglifies, and maps the `JS` files from `/Shared/Scripts` and outputs them to `/wwwroot/Shared/Scripts`. As part of this process, the `package.json` file was rebuilt with only the current dependencies.