Added minimal distribution of *Ext JS 3.2.1*
Unfortunately, there isn't a (reliable) **npm** package source for legacy versions of **Ext JS**. We _may_ update to the current version at some point. For now, however, we continue to use the legacy `3.2.1` version for use of the `TreeView` control. At some point, we need to reevaluate if we want to continue to use **Ext JS** for this (in which case we should upgrade, and use their secure **npm** source) or migrate to an alternate component, since we're no longer using **Ext JS** for any other aspects of the application.