Migrated client-side dependencies to *npm*, located via *Gulp*
Previously, client-side dependencies, such as JavaScript, stylesheets, and fonts, were either manually downloaded or accessed via a CDN. To ensure that the editor is ready for distribution (e.g., via *NuGet*) we're now downloading these via *npm*, and then placing them in their respective directories via the new *Gulp* 'dependency' task. Eventually, we'll likely want to uglify (or minimize) and concat these, so that they don't need to be downloaded independently, but that's not the immediate priority.