Updated to use new custom property mapping
Previously, `AttributeDescriptors` and `PermittedContentTypes` were mapping to their source relationships—`Attributes` and `ContentTypes` respectively. The problem with this is that the `ContentTypeDescriptor` class has a custom implementation of `AttributeDescriptors` and `PermittedContentTypes` that provides a more accurate view. Notably, `AttributeDescriptors` not _only_ includes topics from the `Attributes` list, but _also_ any `topics from _ascendant_ `ContentTypeDescriptor`s, thus allowing attributes to be _inherited_ from _parent_ content types. This ensures all of the correct attributes are displayed on the editor.