Reorganized, commented, and consolidated scripts
Previously, the custom JavaScript files for the site were a bit of a hodgepodge. Functions were repeated between various scripts, varying only slightly. And the names of the scripts didn't necessarily give a good idea of what they were for. To mitigate this, I broke the scripts down by functional area or script, and added comments to further clarify what the functionality is. We'll still want to further clean up and test these; most of the logic was left as-is for this commit.