Implemented `initEditorModal()` popup in `TopicList`
This includes an event listener in the `TreeView`, placeholder markup for the modal in the view, as well as some minor changes to the associated `TopicList.js` to account for changes in the naming convention. Note that while this causes the modal to fire, there remain issues with the actual modal content that need to be addressed. Namely, the close/cancel buttons aren't being displayed, and the navigation isn't being hidden. Those features will need to wait until a future update.