Conditionally hide layout for `IsModal`
There may be a more elegant way of handling this by e.g., conditionally removing the layout—but that introduces other challenges such as having to have an alternate layout without the navigation. For now, this is the easiest, requires the least refactoring, and prevents the need for any dupliction of e.g. script registrations. In addition, (re)introduced classes related to the modal window. Note that while this successfully hides the navigation, it doesn't make the modal window render any better. That's going to require @ktrunkey's attention.