Resolved issues when `TopicListViewComponent` only has one topic
The dropdown list should only be displayed if there are multiple options. Otherwise, a link should be displayed. That said, the options includes an empty-valued label as the first entry. As such, the count will always be one higher than the number of (valid) options. To account for this, the threshold is increased from 1 to 2, and the value and label are pulled from the _last_ entry, not the _first_ entry.