Moved `*DescriptorTopicViewModel` to a `Metadata` namespace
Not only is this consistent with the `Ignia.Topics.Metadata` namespace that these `TopicViewModel`s map to, but it will also make room for the several new `AttributeDescriptor`s that we intend to introduce later today. This required updating the `using` statements in all files that rely on either `ContentTypeDescriptorTopicViewModel` or `AttributeDescriptorTopicViewModel`.