Renamed view components
A number of view components had unintuitive or ambiguous names in order to ensure backward compatibility with the legacy **ASP.NET WebForms** user controls. With the forthcoming change from using an attribute to determine attribute type to a `ContentTypeDescriptor`, we have the opportunity to change these identifiers without too much fuss. For now, since that feature isn't yet complete, the `Edit.cshtml` includes renames to handle the legacy names. As part of this, we're also able to delete intermediary placeholders that were meant to handle existing renames. Note that this commit had to be separated from #1f678f9 because it renamed files to the same name as files that were being deleted or renamed as part of that commit; otherwise, the git history became really confusing, and failed to reflect renames.