Refactored `Options` classes as `AttributeTopicViewModel`s
Previously, for a given view component (e.g., `BooleanViewComponent`) we allows an options object (e.g., `BooleanOptions`) to be constructed and passed along as a set of parameters. These options mapped to options encoded in the `AttributeDescriptorViewModel`'s `DefaultConfiguration` property, which could then be parsed by the view component and set on the options object if a value hadn't already been defined. With updates to **OnTopic**, however, we're now able to separate those out as individual attributes and map them directly to (derivatives of) `AttributeDescriptor`. To facilitate this, the existing `Options` objects are being refactored as `TopicviewModel`s that derive from `AttributeDescriptorTopicViewModel`. This eliminates the ability to easily set configuration options manually on a view component via markup—a feature we rarely need for the editor—but, in turn, gives us strongly typed access to the attribute type configuration.

Technically, we could have both, but this would introduce a lot of duplication as the custom attributes would be defined as properties on both the `TopicViewModel` and the `Options` classes; in the future, we may explore ways of accomplish this by allowing the mapping of e.g., a `TopicViewModel.Options` property. That will be a bigger priority, however, if **Microsoft** exposes the ability to map complex objects to view component attributes (e.g., `<vc:boolean options-is-enabled="true" />`—a capability that's not currently supported, and which makes options objects less optimal for use with view components.

Note: This commit will not compile. Subsequent commits will finish refactoring required to fully incorporate this change.